requirements.txt
Django==1.6.3
django-model-utils==2.0.3
django-grappelli==2.5.3
django-floppyforms==1.2.0
pytz==2014.4

# development dependencies
django-debug-toolbar==1.2
django-extensions==1.3.3
ipdb==0.8
ipython==2.0.0
